+++ library/src/schemas/nullable/nullableAsync.ts
@@ -16,7 +16,9 @@ export type NullableSchemaAsync<
     | Input<TWrapped>
     | (() => Input<TWrapped> | Promise<Input<TWrapped> | undefined> | undefined)
     | undefined = undefined,
-  TOutput = Awaited<TDefault> extends Input<TWrapped>
+  TOutput = TDefault extends
+    | Input<TWrapped>
+    | (() => Input<TWrapped> | Promise<Input<TWrapped>>)
     ? Output<TWrapped>
     : Output<TWrapped> | null
 > = BaseSchemaAsync<Input<TWrapped> | null, TOutput> & {